% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{allClasses}
\alias{allClasses}
\title{Vector of all variable classes in \code{reporteR}}
\usage{
allClasses()
}
\description{
Returns the names of the eight data classes for which
\code{reporteR} is implemented, namely \code{"character"}, \code{"Date"},
\code{"factor"}, \code{"integer"}, \code{"labelled"}, 
\code{"haven_labelled"}, \code{"logical"} and
\code{"numeric"}.
}
\examples{
allClasses()

}
